requirements.txt
flask
flask-sqlalchemy
flask-cors
opencv-python
python-dotenv
pyotp
requests
zeroconf